extract PlayerEventEmitter to separate file
